added orogen plugin for the stream aligner
If the plugin is installed you may use a new orogen syntax for
adding a stream aligner to a task context. A example of the syntax
looks like this :
    add_aggregator() do
        align_port("hbridge_samples", 0.01)
        timeout(0.1)
    end

timout specifies the timeout of the aggregator.
align_port takes the port name and the period of the data as an argument.