timestamper: do the complete estimation without latency
The latency is only added before we return the estimated value to
the caller. The reason is that the estimated latency can vary
(for instance because of estimation errors), and we don't want
latency estimation errors to "leak" into the main estimation loop.

Since latency is a fixed value, ignoring it during estimation has
not effect on the estimation quality.