timestamper: update tests
Turn off drift tests, as they don't make much sense. The timestamp
estimator is based on the idea that you get a periodic stream, which
is not the case if the clock drifts. The tests do pass, though, if
the drift is quite low, but the estimator starts resonating if the
drift is too high.

Fix loss tests: the current tests can pass only if the loss is announced.
We test for both announcement through index and through the explicit call
to updateLoss