timestamper: change how announced losses are handled
Until now, the losses announced to updateLoss were taken "as is",
without cross-referencing with the received time stream.

This changes how they are handled by only "accepting" these losses
if the time stream looks like there is a loss. This adds robustness
to devices that would misreport indexes (as e.g. the hokuyo), and
also makes sure to properly handle situations where we would know
about the loss *before* the sample arrives to the estimator (also
the hokuyo)